Basic tower animation (#76)
* basic animation functionality

* basic animation functionality

* fixed the charge up animation wait time

* provisory attack laser shader

* revert (the attack is now done two times after upgrading- BUG)

* lser fixes(broken animations after upgrading, delta time (<3) and saving the magic prefab updated ref in towerholder prefab instance on screen)

* fix animators

---------

+++ Assets/Scripts/Sheets/TowerSheetNeo.cs
@@ -8,8 +8,8 @@ public class TowerSheetNeo : ScriptableObject
     {
         public int price;
         public int range;
-        public int damage;
-        public int cooldown;
+        public float damage;
+        public float cooldown;
     }
 
     [System.Serializable]
